+++ src/backend/model/gaming/gaming.pri
@@ -2,12 +2,10 @@ HEADERS += \
     $$PWD/Collection.h \
     $$PWD/Game.h \
     $$PWD/GameAssets.h \
-    $$PWD/GameList.h \
     $$PWD/Filters.h \
 
 SOURCES += \
     $$PWD/Collection.cpp \
     $$PWD/Game.cpp \
     $$PWD/GameAssets.cpp \
-    $$PWD/GameList.cpp \
     $$PWD/Filters.cpp \
