--- tests/backend/model/gamelist/gamelist.pro
@@ -1,10 +1,0 @@
-CONFIG += testcase no_testcase_installs
-
-QT += qml testlib
-CONFIG += c++11 warn_on exceptions_off
-
-TARGET = test_GameList
-SOURCES = $${TARGET}.cpp
-DEFINES *= $${COMMON_DEFINES}
-
-include($${TOP_SRCDIR}/src/link_to_backend.pri)
